[IDLE-330] Retorift API Swagger에 분류된 대로 분리
