+++ core/network/src/main/java/com/idle/network/source/auth/AuthDataSource.kt
@@ -1,6 +1,6 @@
 package com.idle.network.source.auth
 
-import com.idle.network.api.CareApi
+import com.idle.network.api.AuthApi
 import com.idle.network.model.auth.BusinessRegistrationResponse
 import com.idle.network.model.auth.ConfirmAuthCodeRequest
 import com.idle.network.model.auth.SendPhoneRequest
@@ -15,41 +15,41 @@ import com.idle.network.util.onResponse
 import javax.inject.Inject
 
 class AuthDataSource @Inject constructor(
-    private val careApi: CareApi
+    private val authApi: AuthApi
 ) {
     suspend fun sendPhoneNumber(sendPhoneRequest: SendPhoneRequest): Result<Unit> =
-        careApi.sendPhoneNumber(sendPhoneRequest).onResponse()
+        authApi.sendPhoneNumber(sendPhoneRequest).onResponse()
 
     suspend fun confirmAuthCode(confirmAuthCodeRequest: ConfirmAuthCodeRequest): Result<Unit> =
-        careApi.confirmAuthCode(confirmAuthCodeRequest).onResponse()
+        authApi.confirmAuthCode(confirmAuthCodeRequest).onResponse()
 
     suspend fun signUpCenter(signUpCenterRequest: SignUpCenterRequest): Result<Unit> =
-        careApi.signUpCenter(signUpCenterRequest).onResponse()
+        authApi.signUpCenter(signUpCenterRequest).onResponse()
 
     suspend fun signInCenter(signInCenterRequest: SignInCenterRequest): Result<TokenResponse> =
-        careApi.signInCenter(signInCenterRequest).onResponse()
+        authApi.signInCenter(signInCenterRequest).onResponse()
 
     suspend fun signUpWorker(signUpWorkerRequest: SignUpWorkerRequest): Result<TokenResponse> =
-        careApi.signUpWorker(signUpWorkerRequest).onResponse()
+        authApi.signUpWorker(signUpWorkerRequest).onResponse()
 
     suspend fun signInWorker(signInWorkerRequest: SignInWorkerRequest): Result<TokenResponse> =
-        careApi.signInWorker(signInWorkerRequest).onResponse()
+        authApi.signInWorker(signInWorkerRequest).onResponse()
 
-    suspend fun logoutWorker(): Result<Unit> = careApi.logoutWorker().onResponse()
+    suspend fun logoutWorker(): Result<Unit> = authApi.logoutWorker().onResponse()
 
-    suspend fun logoutCenter(): Result<Unit> = careApi.logoutCenter().onResponse()
+    suspend fun logoutCenter(): Result<Unit> = authApi.logoutCenter().onResponse()
 
     suspend fun withdrawalCenter(withdrawalCenterRequest: WithdrawalCenterRequest): Result<Unit> =
-        careApi.withdrawalCenter(withdrawalCenterRequest).onResponse()
+        authApi.withdrawalCenter(withdrawalCenterRequest).onResponse()
 
     suspend fun withdrawalWorker(withdrawalWorkerRequest: WithdrawalWorkerRequest): Result<Unit> =
-        careApi.withdrawalWorker(withdrawalWorkerRequest).onResponse()
+        authApi.withdrawalWorker(withdrawalWorkerRequest).onResponse()
 
     suspend fun validateIdentifier(identifier: String): Result<Unit> =
-        careApi.validateIdentifier(identifier).onResponse()
+        authApi.validateIdentifier(identifier).onResponse()
 
     suspend fun validateBusinessRegistrationNumber(
         businessRegistrationNumber: String
     ): Result<BusinessRegistrationResponse> =
-        careApi.validateBusinessRegistrationNumber(businessRegistrationNumber).onResponse()
+        authApi.validateBusinessRegistrationNumber(businessRegistrationNumber).onResponse()
 }
