+++ core/network/src/main/java/com/idle/network/source/jobposting/JobPostingDataSource.kt
@@ -1,6 +1,6 @@
 package com.idle.network.source.jobposting
 
-import com.idle.network.api.CareApi
+import com.idle.network.api.JobPostingApi
 import com.idle.network.model.jobposting.GetCenterJobPostingDetailResponse
 import com.idle.network.model.jobposting.GetJobPostingsInProgressResponse
 import com.idle.network.model.jobposting.GetJobPostingsResponse
@@ -10,30 +10,30 @@ import com.idle.network.util.onResponse
 import javax.inject.Inject
 
 class JobPostingDataSource @Inject constructor(
-    private val careApi: CareApi
+    private val jobPostingApi: JobPostingApi
 ) {
     suspend fun postJobPosting(jobPostingRequest: JobPostingRequest): Result<Unit> =
-        careApi.postJobPosting(jobPostingRequest).onResponse()
+        jobPostingApi.postJobPosting(jobPostingRequest).onResponse()
 
     suspend fun updateJobPosting(
         jobPostingId: String,
         jobPostingRequest: JobPostingRequest
-    ): Result<Unit> = careApi.updateJobPosting(
+    ): Result<Unit> = jobPostingApi.updateJobPosting(
         jobPostingId = jobPostingId,
         jobPostingRequest = jobPostingRequest
     ).onResponse()
 
     suspend fun getCenterJobPostingDetail(jobPostingId: String):
             Result<GetCenterJobPostingDetailResponse> =
-        careApi.getJobPostingDetailCenter(jobPostingId).onResponse()
+        jobPostingApi.getJobPostingDetailCenter(jobPostingId).onResponse()
 
     suspend fun getWorkerJobPostingDetail(jobPostingId: String):
             Result<GetWorkerJobPostingDetailResponse> =
-        careApi.getJobPostingDetailWorker(jobPostingId).onResponse()
+        jobPostingApi.getJobPostingDetailWorker(jobPostingId).onResponse()
 
     suspend fun getJobPostings(next: String?, limit: Int): Result<GetJobPostingsResponse> =
-        careApi.getJobPostings(next = next, limit = limit).onResponse()
+        jobPostingApi.getJobPostings(next = next, limit = limit).onResponse()
 
     suspend fun getJobPostingsInProgress(): Result<GetJobPostingsInProgressResponse> =
-        careApi.getJobPostingsInProgress().onResponse()
+        jobPostingApi.getJobPostingsInProgress().onResponse()
 }
